requirements.txt
# AI-Based Landslide Prediction and Detection System
# Core Dependencies

# Deep Learning & YOLO
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision & Image Processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Data & Visualization
matplotlib>=3.7.0
pandas>=2.0.0
seaborn>=0.12.0

# Web Dashboard
streamlit>=1.28.0

# Satellite Imagery
requests>=2.31.0
sentinelhub>=3.9.0

# Utilities
scipy>=1.10.0
scikit-image>=0.21.0
tqdm>=4.65.0
pyyaml>=6.0
